+++ enhancedcraft-potion-hud/src/main/java/com/spxctreofficial/enhancedcraft/mixin/MinecraftClientMixin.java
@@ -1,0 +1,12 @@
+package com.spxctreofficial.enhancedcraft.mixin;
+
+import net.minecraft.client.MinecraftClient;
+import net.minecraft.client.options.GameOptions;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(MinecraftClient.class)
+public interface MinecraftClientMixin {
+	@Accessor("options")
+	GameOptions getGameOptions();
+}